requirements.txt
Django==5.2.8
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-filter
psycopg2-binary
django-debug-toolbar
python-dotenv
stripe
gunicorn
google-genai
channels
channels-redis
uvicorn
whitenoise
celery[redis]==5.3.6
django-cors-headers